Add a test runner for `kargo test`
This commit adds a test runner when using `kargo test`. Like `cargo`,
this prefers to have very low overhead to getting tests set up, perhaps
at the cost of some deep customizability over how tests are run.

To that end, this commit adds test running for tests using the standard
test annotations and assertions from the `kotlin.test` standard library
using the JUnit5 runner. This means that to run tests, you need a
dependency on `org.jetbrains.kotlin:kotlin-test-junit5`. (TODO: add the
concept of test-only dependencies so that this doesn't end up in the
final jar.) We now vendor the JUnit5 command-line runner at init time and
call out to that for the actual running of tests.

Test discovery / compilation requires that we now care a bit more about
project layout. I'm personally not a fan of the deeply nested
`src/{main,test}/{kotlin,java,scala}` scheme that's the standard and
would prefer to have my test files alongside the code (as is more common
in python, e.g.). In order to allow both, I've added an additional
config option `project_layout` which can be set to `classic` for the
typical `src/main/kotlin`-style layout, or to `flat` for a layout where
the only restriction is that sources need to be in `src` and test files
end in `_test.kt`.

To support these changes and some quick tests in the project, I've
flattened out the project structure, and I've also made it possible to
manually provide a config, which is useful for tests of kargo itself.

Test plan:
- `kargo build`
- `kargo run -- test`