Fix a crash when building a project with no dependencies; test utils
When a project had no dependencies, we'd still pass a bare `-cp`
argument, which would cause the kotlin command-line compiler to error.
This commit fixes it so that we no longer pass any classpath args when
there are no dependencies.

The majority of this commit, however, is adding testing utilities so
that it's possible to have a test for this bug. We now express all paths
relative to a root path (rather than as bare relative paths), which is
set to be the directory where the Kargo.toml file lives. Then, for
testing, we can create a temp dir with a Kargo.toml file and run
commands for testing against that directory.

Fixes #1

Test plan:
- `kargo test`