Add test-only dependencies; automatically include required test dependencies
It saves time and binary size not to include dependencies only needed
for tests to the assembled output jar. This commit adds a
`[test.dependencies]` section to `Kargo.toml` that can be used to
specify test-only dependencies.

The required test-only dependencies for running and annotating tests are
now included automatically and don't need to be listed in Kargo.toml for
every project.

Test plan:
- `kargo test`